% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hide.R
\name{hide}
\alias{hide}
\title{Hide}
\usage{
hide(hidden = c("none"), colour = "white")
}
\arguments{
\item{hidden}{A list that designates what aspect of a figure to conceal. Available options are "all", "xtext", "ytext", "xtitle", "ytitle", "title", "subtitle", "tag", "caption", "xtick" and "ytick". "sub" and "cap" can also be used in place of "subtitle" and "caption" respectively. These options can be added in any order into the hidden list however, the variable must be a list; default is "none"}

\item{colour}{The background color of the figure being created; default is "white"}
}
\value{
A figure with the chosen figure elements hidden without changing the figure dimensions.
}
\description{
Hide
}
\examples{
library("ggplot2")

ggplot(states, aes(x = Pop, y = fct_rev(fct_inorder(State))), fill = NA) +
geom_col(color = "black", fill = "black", width = 0.45, ) +
labs(title = "POPULATION OF EACH STATE",
    subtitle = "(Population in 2021)",
    tag = "tag example",
    caption = "caption example",
    x = "Population",
    y = "State") +
hide("cap", "xtext", "xtitle", "title")

ggplot(states, aes(x = Pop, y = fct_rev(fct_inorder(State))), fill = NA) +
geom_col(color = "black", fill = "black", width = 0.45, ) +
labs(title = "POPULATION OF EACH STATE",
    subtitle = "(Population in 2021)",
    tag = "tag example",
    caption = "caption example",
    x = "Population",
    y = "State") +
hide("caption", "sub", "title", "tag", "xtick", "ytick")

ggplot(states, aes(x = Pop, y = fct_rev(fct_inorder(State))), fill = NA) +
geom_col(color = "black", fill = "black", width = 0.45, ) +
labs(title = "POPULATION OF EACH STATE",
    subtitle = "(Population in 2021)",
    tag = "tag example",
    caption = "caption example",
    x = "Population",
    y = "State") +
hide("all")
}
