% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hide_axistext.R
\name{hide_axistext}
\alias{hide_axistext}
\title{Hide Axis Text}
\usage{
hide_axistext(hide = "both", colour = "white")
}
\arguments{
\item{hide}{The axis you wish to hide; options include "x", "y", and "both"; default is "both"}

\item{colour}{The background color of the figure being created; default is "white"}
}
\value{
A figure with the chosen axis hidden without changing the figure dimensions.
}
\description{
Hide Axis Text
}
\examples{



}
