% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hide_axistitle.R
\name{hide_axistitle}
\alias{hide_axistitle}
\title{Hide Axis Title}
\usage{
hide_axistitle(hide = "both", colour = "white")
}
\arguments{
\item{hide}{The axis title you wish to hide; options include "x", "y", and "both"; default is "both"}

\item{colour}{The background color of the figure being created; default is "white"}
}
\value{
A figure with the chosen axis title hidden without changing the figure dimensions.
}
\description{
Hide Axis Title
}
\examples{
library("ggplot2")

ggplot(states, aes(x = Pop, y = fct_rev(fct_inorder(State))), fill = NA) +
geom_col(color = "black", fill = "black", width = 0.45, ) +
labs(title = "POPULATION OF EACH STATE",
    subtitle = "(Population in 2021)",
    x = "Population",
    y = "State") +
hide _axistitle()

ggplot(states, aes(x = Pop, y = fct_rev(fct_inorder(State))), fill = NA) +
geom_col(color = "black", fill = "black", width = 0.45, ) +
labs(title = "POPULATION OF EACH STATE",
    subtitle = "(Population in 2021)",
    x = "Population",
    y = "State") +
hide _axistitle("x")
}
