% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hide_misctext.R
\name{hide_misctext}
\alias{hide_misctext}
\title{Hide Misc Text}
\usage{
hide_misctext(hide = "both", colour = "white")
}
\arguments{
\item{hide}{Designate whether to hide the plots tag, caption, or both; options include "tag", "cap", and "both"; default is "both"}

\item{colour}{The background color of the figure being created; default is "white"}
}
\value{
A figure with the tag/caption hidden without changing the figure dimensions.
}
\description{
Hide Misc Text
}
\examples{
library("ggplot2")

ggplot(states, aes(x = Pop, y = fct_rev(fct_inorder(State))), fill = NA) +
geom_col(color = "black", fill = "black", width = 0.45, ) +
labs(title = "POPULATION OF EACH STATE",
    subtitle = "(Population in 2021)",
    tag = "tag example",
    caption = "caption example",
    x = "Population",
    y = "State") +
hide_misctext()

ggplot(states, aes(x = Pop, y = fct_rev(fct_inorder(State))), fill = NA) +
geom_col(color = "black", fill = "black", width = 0.45, ) +
labs(title = "POPULATION OF EACH STATE",
    subtitle = "(Population in 2021)",
    tag = "tag example",
    caption = "caption example",
    x = "Population",
    y = "State") +
hide_misctext("cap")
}
