% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_hgp.R
\name{theme_hga}
\alias{theme_hga}
\title{Henry Gannett Atlas Theme}
\usage{
theme_hga(colour = "#f1d9b5")
}
\arguments{
\item{colour}{The background color for the figure; defaults to #f1d9b5}
}
\value{
A figure that resembles the 1890 Statistical Atlas from Henry Gannett.
}
\description{
Henry Gannett Atlas Theme
}
\examples{
library("ggplot2")

ggplot(states, aes(x = Pop, y = fct_rev(fct_inorder(State))), fill = NA) +
geom_col(color = "black", fill = "black", width = 0.45, ) +
labs(title = "POPULATION OF EACH STATE",
    subtitle = "(Population in 2021)",
    x = "Population",
    y = "State") +
theme_hga()
}
