% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_dub.R
\name{theme_dub}
\alias{theme_dub}
\title{Du Bois Theme}
\usage{
theme_dub()
}
\value{
A graph resembles graphs and tables created by Du Bois.
}
\description{
Du Bois Theme
}
\examples{
library("ggplot2")

ggplot(data = Titanic, aes(x = Class, y = Freq, fill = Survived)) +
geom_col() +
theme_dub()
}
