% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_hgp.R
\name{theme_hga}
\alias{theme_hga}
\title{Henry Gannett Atlas Theme}
\usage{
theme_hga()
}
\value{
A graph resembles the 1890 Statistical Atlas from Henry Gannett.
}
\description{
Henry Gannett Atlas Theme
}
\examples{
library("ggplot2")

ggplot(data = Titanic, aes(x = Class, y = Freq, fill = Survived)) +
geom_col() +
theme_hga()
}
